ESP32 refactor, not yet tested
