+++ src/arch/arch.h
@@ -188,7 +188,11 @@ _PM_CUSTOM_BLAST             If defined, instructs core code to not compile
 
 // ARCHITECTURE-SPECIFIC HEADERS -------------------------------------------
 
-#include "esp32.h"
+#include "esp32-common.h"
+#include "esp32.h" // Original ESP32
+#include "esp32-s2.h"
+#include "esp32-s3.h"
+#include "esp32-c3.h"
 #include "nrf52.h"
 #include "rp2040.h"
 #include "samd-common.h"
